+++ src/index.jsx
@@ -2,6 +2,9 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import './services/i18n/i18n';
 import { StoreProvider } from '@ska-telescope/ska-gui-local-storage';
+import { AuthProvider } from '@ska-telescope/ska-login-page';
+import { MSENTRA_CLIENT_ID, MSENTRA_TENANT_ID, MSENTRA_REDIRECT_URI } from './utils/constants';
+
 import App from './components/App/App';
 import Loader from './components/Loader/Loader';
 
@@ -11,7 +14,13 @@ const root = createRoot(container);
 root.render(
   <React.Suspense fallback={<Loader />}>
     <StoreProvider>
-      <App />
+      <AuthProvider
+        MSENTRA_CLIENT_ID={MSENTRA_CLIENT_ID}
+        MSENTRA_TENANT_ID={MSENTRA_TENANT_ID}
+        MSENTRA_REDIRECT_URI={MSENTRA_REDIRECT_URI}
+      >
+        <App />
+      </AuthProvider>
     </StoreProvider>
   </React.Suspense>
 );