+++ webpack.config.js
@@ -2,6 +2,8 @@ const webpack = require('webpack');
 const HtmlWebPackPlugin = require('html-webpack-plugin');
 const ModuleFederationPlugin = require('webpack/lib/container/ModuleFederationPlugin');
 const CopyWebpackPlugin = require('copy-webpack-plugin');
+const Dotenv = require('dotenv-webpack');
+const path = require('path');
 
 const deps = require('./package.json').dependencies;
 
@@ -17,6 +19,11 @@ module.exports = () => {
     },
 
     resolve: {
+      alias: {
+        '@components': path.resolve(__dirname, 'src/components'),
+        '@services': path.resolve(__dirname, 'src/services'),
+        '@utils': path.resolve(__dirname, 'src/utils'),
+      },
       extensions: ['.tsx', '.ts', '.jsx', '.js', '.json'],
     },
 
@@ -111,12 +118,27 @@ module.exports = () => {
             eager: true,
           },
           // SKAO components
+          '@ska-telescope/ska-login-page': {
+            requiredVersion: deps['@ska-telescope/ska-login-page'],
+            eager: true,
+          },
           '@ska-telescope/ska-gui-components': {
             requiredVersion: deps['@ska-telescope/ska-gui-components'],
             eager: true,
           },
           '@ska-telescope/ska-gui-local-storage': {
             requiredVersion: deps['@ska-telescope/ska-gui-local-storage'],
+            eager: true,
+          },
+          // MS Entra components
+          '@azure/msal-browser': {
+            requiredVersion: deps['@azure/msal-browser'],
+            singleton: true,
+            eager: true,
+          },
+          '@azure/msal-react': {
+            requiredVersion: deps['@azure/msal-react'],
+            singleton: true,
             eager: true,
           },
           // mixture
@@ -138,6 +160,7 @@ module.exports = () => {
         },
       }),
       new HtmlWebPackPlugin({
+        inject: true,
         template: './public/index.html',
       }),
       new CopyWebpackPlugin({
@@ -151,6 +174,9 @@ module.exports = () => {
             },
           },
         ],
+      }),
+      new Dotenv({
+        path: '.env',
       }),
     ],
   };