+++ CHANGELOG.md
@@ -7,7 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 1.1.0
 
-Addition of the ska-login components as a working example
+Addition of the ska-login components as a working example including annotation in the codebase
+Some small documentation updates
 
 1.0.3